requirements.txt
apache-airflow-providers-google>=15.1.0
apache-airflow-providers-postgres>=6.1.3
apache-airflow-providers-celery>=3.10.6
google-cloud-storage>=2.7.0
google-cloud-bigquery>=3.31.0
google-cloud-bigquery-connection>=1.18.2
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyarrow>=14.0.1
